--- api/src/azure-blob-storage/core/application/azure-blob-storage.application.ts
@@ -1,5 +1,0 @@
-import { UploadFileToAzureStorageCommandHandler } from './commands/upload-file-to-azure-storage/upload-file-to-azure-storage.command-handler';
-
-const AzureBlobStorageCommandHandler = [UploadFileToAzureStorageCommandHandler];
-
-export const AzureBlobStorageApplications = [...AzureBlobStorageCommandHandler];
