+++ api/src/azure-blob-storage/core/application/ports/azure-blob-storage.port.ts
@@ -1,3 +1,4 @@
+import 'multer';
 export abstract class AzureBlobStoragePort {
   abstract upload(
     file: Express.Multer.File,
