+++ api/src/azure-blob-storage/infrastructure/azure-blob-storage.adapter.ts
@@ -9,20 +9,19 @@ export class AzureBlobStorageAdapter implements AzureBlobStoragePort {
   constructor(private readonly configService: ConfigService) {}
   private readonly logger = new Logger();
 
+  AZURE_STORAGE_URL = this.configService.get<string>('AZURE_STORAGE_URL', '');
+
   private connect(containerName: string): ContainerClient {
-    this.logger.log(`AzureBlobStorageAdapter > connect > start`);
-    const AZURE_STORAGE_URL = this.configService.get<string>(
-      'AZURE_STORAGE_URL',
-      '',
-    );
-    if (!AZURE_STORAGE_URL) throw new Error(AZURE_BLOB_STORAGE_NO_CREDENTIALS);
+    if (!this.AZURE_STORAGE_URL)
+      throw new Error(AZURE_BLOB_STORAGE_NO_CREDENTIALS);
 
     const ENV = this.configService.get<string>('ENV', 'development');
     if (ENV === TEST_ENVIRONMENT) {
       containerName = `${containerName}-test`;
     }
-    const soundGifAzureBlobStorage =
-      BlobServiceClient.fromConnectionString(AZURE_STORAGE_URL);
+    const soundGifAzureBlobStorage = BlobServiceClient.fromConnectionString(
+      this.AZURE_STORAGE_URL,
+    );
     const containerClient =
       soundGifAzureBlobStorage.getContainerClient(containerName);
     return containerClient;
@@ -36,8 +35,8 @@ export class AzureBlobStorageAdapter implements AzureBlobStoragePort {
     this.logger.log(
       `AzureBlobStorageAdapter > upload > called with fileName: ${fileName} and containerName: ${containerName}`,
     );
-    const blob = new Blob([file.buffer]);
     const containerClient = this.connect(containerName);
+    const blob = new Blob([file.buffer]);
     const blockBlobClient = containerClient.getBlockBlobClient(fileName);
     await blockBlobClient.uploadData(blob);
     return blockBlobClient.url;
