--- api/src/azure-blob-storage/interface/azure-blob-storage.interface.ts
@@ -1,5 +1,0 @@
-import { AzureBlobStoragePresenter } from './azure-blob-storage.presenter';
-
-export const AzureBlobStorageInterface = {
-  presenters: [AzureBlobStoragePresenter],
-};
