+++ api/src/sound-gif/core/application/commands/create-sound-gif/create-sound-gif.command-handler.spec.ts
@@ -1,14 +1,10 @@
 import { ConfigService } from '@nestjs/config';
 import { Blob } from 'buffer';
-import { AzureBlobStoragePresenter } from '../../../../../azure-blob-storage/interface/azure-blob-storage.presenter';
+import * as faker from 'faker';
 import { CreateSoundGifCommand } from './create-sound-gif.command';
 import { CreateSoundGifCommandHandler } from './create-sound-gif.command-handler';
 
 const configService = new ConfigService();
-
-const azureStoragePresenter = {
-  upload: jest.fn(),
-};
 
 const createSoundGifPort = {
   create: jest.fn(),
@@ -21,8 +17,8 @@ blob['name'] = title;
 
 const payload = {
   title,
-  audioFile: blob as unknown as Express.Multer.File,
-  imageFile: blob as unknown as Express.Multer.File,
+  audioUrl: faker.internet.url(),
+  imageUrl: faker.internet.url(),
   containerName: 'images',
 };
 
@@ -34,7 +30,6 @@ describe('createSoundGifCommand', () => {
   const createSoundGifCommandHandler = new CreateSoundGifCommandHandler(
     configService,
     createSoundGifPort,
-    azureStoragePresenter as unknown as AzureBlobStoragePresenter,
   );
 
   it('should create a soundgif', async () => {
