+++ api/src/sound-gif/interface/sound-gif.interface.ts
@@ -1,6 +1,6 @@
-import { SoundGifController } from './sound-gif.controller';
+import { CreateSoundGifController } from './controllers/create-sound-gif.controllers';
+import { FindSoundGifController } from './controllers/find-sound-gif.controller';
 
 export const SoundGifInterface = {
-  resolvers: [],
-  controllers: [SoundGifController],
+  controllers: [FindSoundGifController, CreateSoundGifController],
 };
